Extract config to separate file, add helpers
